requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
numpy
python-multipart
httpx
locust
streamlit
plotly
pandas
psutil
requests
pytest
pytest-asyncio